- Refactored Leads to Sales Opportunities and Contacts to leads - Modified Sales Opportunity form - Modified Routes and menu item names - moved lead excel upload to leads list
