--- client-app/src/features/CRM/contacts/ContactsDashboard.tsx
@@ -1,95 +1,0 @@
-import React, { useState, useCallback } from 'react';
-import { Paper, Box, Typography } from '@mui/material';
-import Button from '@mui/material/Button';
-import AddIcon from '@mui/icons-material/Add';
-import { useTranslationHelper } from '../../../app/hooks/useTranslationHelper';
-import CRMMenu from '../menu/CRMMenu';
-import ContactsList from './ContactsList';
-import ContactForm from './ContactForm';
-import { Contact } from '../models/contact';
-
-type EditMode = 'none' | 'create' | 'edit';
-
-const ContactsDashboard: React.FC = () => {
-    const { getTranslatedLabel } = useTranslationHelper();
-    const localizationKey = 'crm.contacts';
-
-    const [editMode, setEditMode] = useState<number>(0);
-    const [selectedContact, setSelectedContact] = useState<Contact | undefined>();
-
-    const handleCreateNew = useCallback(() => {
-        setSelectedContact(undefined);
-        setEditMode(1);
-    }, []);
-
-    const handleEditContact = useCallback((contact: Contact) => {
-        setSelectedContact(contact);
-        setEditMode(2);
-    }, []);
-
-    const handleCloseForm = useCallback(() => {
-        setEditMode(0);
-        setSelectedContact(undefined);
-    }, []);
-
-    const handleFormSuccess = useCallback(() => {
-        // RTK Query cache invalidation will refresh the data
-    }, []);
-
-    return (
-        <>
-            <CRMMenu selectedMenuItem="contacts" />
-
-            <Paper
-                elevation={5}
-                className="div-container-withBorderCurved"
-                sx={{ mt: 2, mx: 2 }}
-            >
-                {/* Header */}
-                <Box
-                    sx={{
-                        display: 'flex',
-                        justifyContent: 'space-between',
-                        alignItems: 'center',
-                        p: 2,
-                        borderBottom: '1px solid',
-                        borderColor: 'divider'
-                    }}
-                >
-                    <Typography variant="h5" fontWeight="medium">
-                        {getTranslatedLabel(`${localizationKey}.title`, 'People')}
-                    </Typography>
-
-                    {editMode === 0 && (
-                        <Button
-                            variant="contained"
-                            color="primary"
-                            sx={{ gap: 1 }}
-                            onClick={handleCreateNew}
-                        >
-                            <AddIcon />
-                            {getTranslatedLabel(`${localizationKey}.createNew`, 'New Contact')}
-                        </Button>
-                    )}
-                </Box>
-
-                {/* Content */}
-                {editMode !== 0 ? (
-                    <ContactForm
-                        contact={selectedContact}
-                        editMode={editMode}
-                        onClose={handleCloseForm}
-                        onSuccess={handleFormSuccess}
-                    />
-                ) : (
-                    <ContactsList
-                        onCreateNew={handleCreateNew}
-                        onEditContact={handleEditContact}
-                    />
-                )}
-            </Paper>
-        </>
-    );
-};
-
-export default ContactsDashboard;
