+++ client-app/src/features/CRM/leads/LeadsDashboard.tsx
@@ -1,80 +1,66 @@
 import React, { useState, useCallback } from 'react';
-import {
-    Paper,
-    Box,
-    ToggleButtonGroup,
-    ToggleButton,
-    Typography
-} from '@mui/material';
+import { Paper, Box, Typography } from '@mui/material';
 import Button from '@mui/material/Button';
-import ViewListIcon from '@mui/icons-material/ViewList';
-import ViewKanbanIcon from '@mui/icons-material/ViewKanban';
 import AddIcon from '@mui/icons-material/Add';
+import DriveFolderUploadIcon from '@mui/icons-material/DriveFolderUpload';
 import { useTranslationHelper } from '../../../app/hooks/useTranslationHelper';
 import CRMMenu from '../menu/CRMMenu';
-import LeadsBoard from './LeadsBoard';
 import LeadsList from './LeadsList';
-import LeadForm from './LeadForm';
-import { SalesOpportunity } from '../models/salesOpportunity';
-import ImportedDataGrid from './ImportedDataGrid';
-import ExcelUploadDialog from './ExcelUploadDialog';
+import LeadsForm from './LeadsForm'
+import { Contact } from '../models/contact';
+import ExcelUploadDialog from './ExcelUploadDialog'
+import ImportedDataGrid from './ImportedDataGrid'
 
-type ViewMode = 'board' | 'list';
 type EditMode = 'none' | 'create' | 'edit';
 
 const LeadsDashboard: React.FC = () => {
     const { getTranslatedLabel } = useTranslationHelper();
-    const localizationKey = 'crm.leads';
+    const localizationKey = 'crm.contacts';
 
-    const [viewMode, setViewMode] = useState<ViewMode>('board');
-    const [editMode, setEditMode] = useState<EditMode>('none');
-    const [selectedOpportunity, setSelectedOpportunity] = useState<SalesOpportunity | undefined>();
+    const [editMode, setEditMode] = useState<number>(0);
+    const [selectedLead, setSelectedLead] = useState<Contact | undefined>();
     const [uploadOpen, setUploadOpen] = useState(false);
-const [importedData, setImportedData] = useState<any[]>([]);
-const [importedFileName, setImportedFileName] = useState('');
-const [showImportedGrid, setShowImportedGrid] = useState(false);
+    const [importedData, setImportedData] = useState<any[]>([]);
+    const [importedFileName, setImportedFileName] = useState('');
+    const [showImportedGrid, setShowImportedGrid] = useState(false);
 
-const handleUploadExcel = () => {
-  setUploadOpen(true);
-};
+    const handleUploadExcel = () => {
+        setUploadOpen(true);
+    };
 
-const handleDataParsed = (data: any[], fileName: string) => {
-  setImportedData(data);
-  setImportedFileName(fileName);
-  setShowImportedGrid(true);
-  setUploadOpen(false);
-};
+    const handleDataParsed = (data: any[], fileName: string) => {
+        setImportedData(data);
+        setImportedFileName(fileName);
+        setShowImportedGrid(true);
+        setUploadOpen(false);
+    };
 
-const handleCloseImported = () => {
-  setShowImportedGrid(false);
-  setImportedData([]);
-  setImportedFileName('');
-};
-
-    const handleViewChange = useCallback((_: React.MouseEvent<HTMLElement>, newView: ViewMode | null) => {
-        if (newView !== null) {
-            setViewMode(newView);
-        }
-    }, []);
+    const handleCloseImported = () => {
+        setShowImportedGrid(false);
+        setImportedData([]);
+        setImportedFileName('');
+    };
 
     const handleCreateNew = useCallback(() => {
-        setSelectedOpportunity(undefined);
-        setEditMode('create');
+        setSelectedLead(undefined);
+        setEditMode(1);
     }, []);
 
-    const handleEditOpportunity = useCallback((opportunity: SalesOpportunity) => {
-        setSelectedOpportunity(opportunity);
-        setEditMode('edit');
+    const handleEditContact = useCallback((lead: Contact) => {
+        setSelectedLead(lead);
+        setEditMode(2);
     }, []);
 
     const handleCloseForm = useCallback(() => {
-        setEditMode('none');
-        setSelectedOpportunity(undefined);
+        setEditMode(0);
+        setSelectedLead(undefined);
     }, []);
 
     const handleFormSuccess = useCallback(() => {
-        // The RTK Query cache invalidation will automatically refresh the data
+        // RTK Query cache invalidation will refresh the data
     }, []);
+
+    const handleCreateBatchLeads = () => {}
 
     return (
         <>
@@ -85,7 +71,7 @@ const handleCloseImported = () => {
                 className="div-container-withBorderCurved"
                 sx={{ mt: 2, mx: 2 }}
             >
-                {/* Header with View Switcher */}
+                {/* Header */}
                 <Box
                     sx={{
                         display: 'flex',
@@ -97,83 +83,77 @@ const handleCloseImported = () => {
                     }}
                 >
                     <Typography variant="h5" fontWeight="medium">
-                        {getTranslatedLabel(`${localizationKey}.title`, 'Sales Pipeline')}
+                        {getTranslatedLabel(`${localizationKey}.title`, 'Leads')}
                     </Typography>
 
-                    <Box sx={{ display: 'flex', gap: 2, alignItems: 'center' }}>
-                        {/* View Toggle */}
-                        <ToggleButtonGroup
-                            value={viewMode}
-                            exclusive
-                            onChange={handleViewChange}
-                            size="small"
-                        >
-                            <ToggleButton value="board" aria-label="board view" sx={{ gap: 1}}>
-                                <ViewKanbanIcon />
-                                {getTranslatedLabel(`${localizationKey}.boardView`, 'Board')}
-                            </ToggleButton>
-                            <ToggleButton value="list" aria-label="list view" sx={{ gap: 1}}>
-                                <ViewListIcon />
-                                {getTranslatedLabel(`${localizationKey}.listView`, 'List')}
-                            </ToggleButton>
-                        </ToggleButtonGroup>
-
-                        {/* Create Button (shown in board view) */}
-                        {viewMode === 'board' && (
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                sx={{ gap: 1 }}
-                                onClick={handleCreateNew}
-                            >
-                                <AddIcon />
-                                {getTranslatedLabel(`${localizationKey}.createNew`, 'New Opportunity')}
-                            </Button>
-                        )}
-                        {viewMode === 'board' && (
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                sx={{ gap: 1 }}
-                                onClick={handleUploadExcel}
-                            >
-                                <AddIcon />
-                                {getTranslatedLabel(`${localizationKey}.uploadExcel`, 'Upload Excel')}
-                            </Button>
-                        )}
-                    </Box>
+                    {editMode === 0 && (
+                        <Box sx={{
+                            display: 'flex',
+                            alignItems: 'center',
+                            gap: 2
+                        }}>
+                            {!showImportedGrid && <>
+                               <Button
+                                    variant="contained"
+                                    color="primary"
+                                    sx={{ gap: 1 }}
+                                    onClick={handleCreateNew}
+                                >
+                                    <AddIcon />
+                                    {getTranslatedLabel(`${localizationKey}.createNew`, 'New Contact')}
+                                </Button>
+                                <Button
+                                    variant="contained"
+                                    color="secondary"
+                                    sx={{ gap: 1 }}
+                                    onClick={handleUploadExcel}
+                                >
+                                    <DriveFolderUploadIcon />
+                                    {getTranslatedLabel(`${localizationKey}.uploadExcel`, 'Upload Excel')}
+                                </Button>
+                            </>}
+                            {showImportedGrid && (
+                                <Button
+                                    variant="contained"
+                                    color="primary"
+                                    sx={{ gap: 1 }}
+                                    onClick={handleCreateBatchLeads}
+                                >
+                                    <DriveFolderUploadIcon />
+                                    {getTranslatedLabel(`${localizationKey}.submitLeads`, 'Create Leads')}
+                                </Button>
+                            )}
+                        </Box>
+                    )}
                 </Box>
 
-                {/* Content Area */}
-                {editMode !== 'none' ? (
-                    <LeadForm
-                        opportunity={selectedOpportunity}
+                {/* Content */}
+                {showImportedGrid ? (
+                    <ImportedDataGrid
+                        data={importedData}
+                        fileName={importedFileName}
+                        onClose={handleCloseImported}
+                    />
+                ) : (editMode !== 0) ? (
+                    <LeadsForm
+                        lead={selectedLead}
                         editMode={editMode}
                         onClose={handleCloseForm}
                         onSuccess={handleFormSuccess}
                     />
-                ) : viewMode === 'board' ? (
-  showImportedGrid ? (
-    <ImportedDataGrid
-      data={importedData}
-      fileName={importedFileName}
-      onClose={handleCloseImported}
-    />
-  ) : (
-    <LeadsBoard onEditOpportunity={handleEditOpportunity} />
-  )
-): (
+                ) : (
                     <LeadsList
                         onCreateNew={handleCreateNew}
-                        onEditOpportunity={handleEditOpportunity}
+                        onEditContact={handleEditContact}
                     />
                 )}
             </Paper>
             <ExcelUploadDialog
-  open={uploadOpen}
-  onClose={() => setUploadOpen(false)}
-  onDataParsed={handleDataParsed}
-/>
+                open={uploadOpen}
+                onClose={() => setUploadOpen(false)}
+                onDataParsed={handleDataParsed}
+            />
+            
         </>
     );
 };
