+++ client-app/src/features/CRM/leads/LeadsList.tsx
@@ -3,118 +3,105 @@ import {
     Grid as KendoGrid,
     GridColumn as Column,
     GridToolbar,
-    GRID_COL_INDEX_ATTRIBUTE,
     GridDataStateChangeEvent,
     GridCellProps
 } from '@progress/kendo-react-grid';
 import { State, process } from '@progress/kendo-data-query';
-import { useTableKeyboardNavigation } from '@progress/kendo-react-data-tools';
-import { Grid, Box, Chip, Typography } from '@mui/material';
+import { Grid, Typography, Chip, Box } from '@mui/material';
 import Button from '@mui/material/Button';
 import { useTranslationHelper } from '../../../app/hooks/useTranslationHelper';
-import { useFetchOpportunitiesQuery } from '../../../app/store/configureStore';
-import { SalesOpportunity } from '../models/salesOpportunity';
+import { useFetchContactsQuery } from '../../../app/store/configureStore';
+import { Contact } from '../models/contact';
 import LoadingComponent from '../../../app/layout/LoadingComponent';
 
-interface LeadsListProps {
+interface ContactsListProps {
     onCreateNew: () => void;
-    onEditOpportunity: (opportunity: SalesOpportunity) => void;
+    onEditContact: (contact: Contact) => void;
 }
 
-const LeadsList: React.FC<LeadsListProps> = ({ onCreateNew, onEditOpportunity }) => {
+const LeadsList: React.FC<ContactsListProps> = ({ onCreateNew, onEditContact }) => {
     const { getTranslatedLabel } = useTranslationHelper();
-    const localizationKey = 'crm.leads.list';
+    const localizationKey = 'crm.contacts.list';
 
-    const { data: opportunities, isLoading, refetch } = useFetchOpportunitiesQuery();
+    const { data: contacts, isLoading } = useFetchContactsQuery();
 
     const [dataState, setDataState] = React.useState<State>({
         take: 10,
         skip: 0,
-        sort: [{ field: 'createdStamp', dir: 'desc' }]
+        sort: [{ field: 'fullName', dir: 'asc' }]
     });
 
     const dataStateChange = (e: GridDataStateChangeEvent) => {
         setDataState(e.dataState);
     };
 
-    const processedData = opportunities
-        ? process(opportunities, dataState)
+    const processedData = contacts
+        ? process(contacts, dataState)
         : { data: [], total: 0 };
 
-    // Custom cell for opportunity name (clickable)
+    // Custom cell for name (clickable)
     const NameCell = (props: GridCellProps) => {
-        const navigationAttributes = useTableKeyboardNavigation(props.id);
         return (
-            <td
-                className={props.className}
-                style={{ ...props.style }}
-                colSpan={props.colSpan}
-                role="gridcell"
-                aria-colindex={props.ariaColumnIndex}
-                aria-selected={props.isSelected}
-                {...{ [GRID_COL_INDEX_ATTRIBUTE]: props.columnIndex }}
-                {...navigationAttributes}
-            >
+            <td className={props.className} style={props.style}>
                 <Button
                     variant="text"
                     color="primary"
-                    onClick={() => onEditOpportunity(props.dataItem)}
+                    onClick={() => onEditContact(props.dataItem)}
                     sx={{ textTransform: 'none', justifyContent: 'flex-start', p: 0 }}
                 >
-                    {props.dataItem.opportunityName}
+                    {props.dataItem.fullName || `${props.dataItem.firstName} ${props.dataItem.lastName}`.trim()}
                 </Button>
             </td>
         );
     };
 
-    // Custom cell for amount
-    const AmountCell = (props: GridCellProps) => {
-        const amount = props.dataItem.estimatedAmount || 0;
-        const currency = props.dataItem.currencyUomId || 'USD';
-        const formatted = new Intl.NumberFormat('en-US', {
-            style: 'currency',
-            currency,
-            minimumFractionDigits: 0
-        }).format(amount);
-
+    // Custom cell for email
+    const EmailCell = (props: GridCellProps) => {
+        const email = props.dataItem.email;
         return (
             <td className={props.className} style={props.style}>
-                <Typography variant="body2" fontWeight="medium" color="success.main">
-                    {formatted}
-                </Typography>
+                {email ? (
+                    <a href={`mailto:${email}`} style={{ color: 'inherit' }}>
+                        {email}
+                    </a>
+                ) : (
+                    <Typography variant="caption" color="text.disabled">-</Typography>
+                )}
             </td>
         );
     };
 
-    // Custom cell for stage
-    const StageCell = (props: GridCellProps) => {
+    // Custom cell for phone
+    const PhoneCell = (props: GridCellProps) => {
+        const phone = props.dataItem.phone || props.dataItem.mobilePhone;
         return (
             <td className={props.className} style={props.style}>
-                <Chip
-                    label={props.dataItem.opportunityStageName || props.dataItem.opportunityStageId}
-                    size="small"
-                    color="primary"
-                    variant="outlined"
-                />
+                {phone ? (
+                    <a href={`tel:${phone}`} style={{ color: 'inherit' }}>
+                        {phone}
+                    </a>
+                ) : (
+                    <Typography variant="caption" color="text.disabled">-</Typography>
+                )}
             </td>
         );
     };
 
-    // Custom cell for probability
-    const ProbabilityCell = (props: GridCellProps) => {
-        const probability = props.dataItem.estimatedProbability || 0;
-        let color: 'error' | 'warning' | 'success' = 'error';
-        if (probability >= 70) color = 'success';
-        else if (probability >= 40) color = 'warning';
-
+    // Custom cell for status
+    const StatusCell = (props: GridCellProps) => {
+        const status = props.dataItem.statusDescription || props.dataItem.statusId;
         return (
             <td className={props.className} style={props.style}>
-                <Chip
-                    label={`${probability}%`}
-                    size="small"
-                    color={color}
-                    sx={{ minWidth: 50 }}
-                />
+                {status ? (
+                    <Chip
+                        label={status}
+                        size="small"
+                        color={props.dataItem.statusId === 'PARTY_ENABLED' ? 'success' : 'default'}
+                        variant="outlined"
+                    />
+                ) : (
+                    <Typography variant="caption" color="text.disabled">-</Typography>
+                )}
             </td>
         );
     };
@@ -132,30 +119,17 @@ const LeadsList: React.FC<LeadsListProps> = ({ onCreateNew, onEditOpportunity })
         );
     };
 
-    // Custom cell for contacts
-    const ContactsCell = (props: GridCellProps) => {
-        const contacts = props.dataItem.contacts || [];
+    // Custom cell for address
+    const AddressCell = (props: GridCellProps) => {
+        const { address1, city, countryGeoId } = props.dataItem;
+        const parts = [address1, city, countryGeoId].filter(Boolean);
+
         return (
             <td className={props.className} style={props.style}>
-                {contacts.length > 0 ? (
-                    <Box sx={{ display: 'flex', gap: 0.5, flexWrap: 'wrap' }}>
-                        {contacts.slice(0, 2).map((c: any, idx: number) => (
-                            <Chip
-                                key={idx}
-                                label={c.partyName}
-                                size="small"
-                                variant="outlined"
-                                sx={{ maxWidth: 100 }}
-                            />
-                        ))}
-                        {contacts.length > 2 && (
-                            <Chip
-                                label={`+${contacts.length - 2}`}
-                                size="small"
-                                color="default"
-                            />
-                        )}
-                    </Box>
+                {parts.length > 0 ? (
+                    <Typography variant="body2" noWrap sx={{ maxWidth: 200 }}>
+                        {parts.join(', ')}
+                    </Typography>
                 ) : (
                     <Typography variant="caption" color="text.disabled">-</Typography>
                 )}
@@ -164,7 +138,7 @@ const LeadsList: React.FC<LeadsListProps> = ({ onCreateNew, onEditOpportunity })
     };
 
     if (isLoading) {
-        return <LoadingComponent message={getTranslatedLabel(`${localizationKey}.loading`, 'Loading opportunities...')} />;
+        return <LoadingComponent message={getTranslatedLabel(`${localizationKey}.loading`, 'Loading contacts...')} />;
     }
 
     return (
@@ -186,64 +160,41 @@ const LeadsList: React.FC<LeadsListProps> = ({ onCreateNew, onEditOpportunity })
                     <Grid container justifyContent="space-between" alignItems="center">
                         <Grid item>
                             <Typography variant="h6">
-                                {getTranslatedLabel(`${localizationKey}.title`, 'Sales Opportunities')}
+                                {getTranslatedLabel(`${localizationKey}.title`, 'Contacts')}
                             </Typography>
-                        </Grid>
-                        <Grid item>
-                            <Button
-                                variant="outlined"
-                                color="secondary"
-                                onClick={onCreateNew}
-                            >
-                                {getTranslatedLabel(`${localizationKey}.createNew`, 'Create New Opportunity')}
-                            </Button>
                         </Grid>
                     </Grid>
                 </GridToolbar>
 
                 <Column
-                    field="opportunityName"
+                    field="fullName"
                     title={getTranslatedLabel(`${localizationKey}.name`, 'Name')}
                     cell={NameCell}
-                    width={250}
+                    width={200}
                     locked={true}
                 />
                 <Column
-                    field="estimatedAmount"
-                    title={getTranslatedLabel(`${localizationKey}.amount`, 'Value')}
-                    cell={AmountCell}
-                    width={120}
+                    field="email"
+                    title={getTranslatedLabel(`${localizationKey}.email`, 'Email')}
+                    cell={EmailCell}
                 />
                 <Column
-                    field="opportunityStageName"
-                    title={getTranslatedLabel(`${localizationKey}.stage`, 'Stage')}
-                    cell={StageCell}
-                    width={130}
-                />
-                <Column
-                    field="estimatedProbability"
-                    title={getTranslatedLabel(`${localizationKey}.probability`, 'Probability')}
-                    cell={ProbabilityCell}
-                    width={100}
-                />
-                <Column
-                    field="ownerName"
-                    title={getTranslatedLabel(`${localizationKey}.owner`, 'Owner')}
+                    field="phone"
+                    title={getTranslatedLabel(`${localizationKey}.phone`, 'Phone')}
+                    cell={PhoneCell}
                     width={150}
                 />
                 <Column
-                    field="estimatedCloseDate"
-                    title={getTranslatedLabel(`${localizationKey}.closeDate`, 'Close Date')}
-                    cell={DateCell}
-                    width={120}
+                    field="address1"
+                    title={getTranslatedLabel(`${localizationKey}.address`, 'Address')}
+                    cell={AddressCell}
+                    width={250}
                 />
                 <Column
-                    field="contacts"
-                    title={getTranslatedLabel(`${localizationKey}.contacts`, 'Contacts')}
-                    cell={ContactsCell}
-                    width={200}
-                    filterable={false}
-                    sortable={false}
+                    field="statusDescription"
+                    title={getTranslatedLabel(`${localizationKey}.status`, 'Status')}
+                    cell={StatusCell}
+                    width={120}
                 />
                 <Column
                     field="createdStamp"
