+++ client-app/src/features/CRM/sales-opportunities/SalesOpportunityBoard.tsx
